Added comments, replaced code to /src, added config and .env file
