+++ src/config/.env.sample
@@ -1,0 +1,1 @@
+DB_URL =
