+++ src/config/config.ts
@@ -1,0 +1,10 @@
+import * as dotenv from "dotenv"
+
+dotenv.config()
+
+export class Config {
+    /**
+     * Database URL
+     */
+    public static dbUrl: string = process.env.DB_URL || '';
+}
